make pseudofiles static and single build
rename and format

pseudofiles: remove unused imports

update naming

fix up read bytes
